chore: clean up unused files and rename component
