--- src/views/AboutUsPage/AboutUs.css
@@ -1,262 +1,0 @@
-.w-fit {
-  width: fit-content;
-}
-
-.w-full {
-  width: 100%;
-}
-
-.max-w-999px {
-  width: 100%;
-  max-width: 999px;
-}
-
-.mx-auto {
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 0;
-  margin-bottom: 0;
-}
-
-.px-22px {
-  padding: 0 22px;
-}
-
-.mb-auto {
-  margin-bottom: auto;
-}
-
-.grid {
-  display: grid;
-}
-
-.grid-cols-1fr_auto {
-  grid-template-columns: auto 1fr;
-}
-
-.me-3px {
-  margin-right: 3px;
-}
-
-.vertical-text {
-  writing-mode: vertical-lr;
-  transform: rotate(180deg);
-}
-
-@media (max-width: 960px) {
-  .grid-cols-1fr_auto {
-    grid-template-columns: 1fr;
-    gap: 1.25rem;
-  }
-
-  .vertical-text {
-    writing-mode: horizontal-tb;
-    transform: rotate(0deg);
-  }
-
-  .mb-auto {
-    margin-bottom: 0;
-  }
-}
-
-.title {
-  font-size: 1.5rem;
-  font-weight: 700;
-  color: #1d3967;
-  line-height: 150%;
-}
-
-.paragraphs {
-  font-size: 1.5rem;
-  font-weight: 300;
-  color: #1d3967;
-  line-height: 130%;
-  letter-spacing: 0.5px;
-  padding: 0 1.25rem;
-}
-
-@media (max-width: 601px) {
-  .paragraphs {
-    font-size: 1.3rem;
-    line-height: 140%;
-    padding: 0 0rem;
-  }
-}
-
-.text-center {
-  text-align: center;
-}
-
-/* ImageGallery.css */
-/* ImageGallery.css */
-
-/* Gallery container */
-.img-collection {
-  position: relative;
-  display: grid;
-  width: 100%;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 1rem;
-}
-
-/* Tablet: 2 columns */
-@media (max-width: 1279px) {
-  .img-collection {
-    grid-template-columns: repeat(2, 1fr);
-    /* 2 columns */
-  }
-}
-
-/* Phone: 1 column */
-@media (max-width: 601px) {
-  .img-collection {
-    grid-template-columns: repeat(1, 1fr);
-    /* 1 column */
-  }
-}
-
-/* Image container */
-.img-wrapper {
-  position: relative;
-  display: block;
-  width: 100%;
-  height: auto;
-  overflow: hidden;
-  grid-column: span 1 / span 1;
-}
-
-/* Individual image */
-.img {
-  display: block;
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  transform: scale(1.02);
-}
-
-/* Navigation arrows */
-.navigation-arrows {
-  width: fit-content;
-  display: flex;
-  gap: 2rem;
-  margin: 20px 0 30px auto;
-}
-
-@media (max-width: 601px) {
-  .navigation-arrows {
-    margin: 20px 0 20px auto;
-  }
-}
-
-.arrow {
-  cursor: pointer;
-  padding: 0 0.5rem;
-  color: #1d3967;
-  font-size: 24px;
-  -webkit-user-select: none;
-  user-select: none;
-}
-
-/* Responsive behavior */
-@media (max-width: 768px) {
-  .gallery-container {
-    width: 100%;
-  }
-
-  .gallery-image {
-    width: 100%;
-  }
-}
-
-.quote {
-  position: relative;
-  margin: 0px 0 20px 0;
-  width: fit-content;
-  transform: rotate(0deg);
-}
-
-.rotate-180 {
-  transform: rotate(180deg);
-  margin: 10px 0 0px auto;
-}
-
-.socials {
-  display: flex;
-  flex-direction: column;
-  gap: 27px;
-  width: fit-content;
-  margin-right: 20px;
-}
-
-@media (max-width: 960px) {
-  .socials {
-    display: none;
-  }
-}
-
-.h-50px {
-  height: 50px;
-}
-
-.h-105px {
-  height: 105px;
-}
-
-.min-768px-px-1rem {
-  padding: 0 0.5rem;
-}
-
-@media (max-width: 768px) {
-  .min-768px-px-1rem {
-    padding: 0 0rem;
-  }
-}
-
-.flex-between {
-  display: flex;
-  justify-content: space-between;
-  gap: 0.5rem;
-  flex-wrap: wrap;
-}
-
-.flex-column {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  width: 210px;
-}
-@media (max-width: 500px) {
-  .flex-between {
-    gap: 1.5rem;
-  }
-  .flex-column {
-    width: 170px;
-  }
-}
-
-.sm-hidden,
-.md-hidden {
-  display: none;
-}
-@media (max-width: 960px) {
-  .md-hidden {
-    display: flex;
-  }
-}
-@media (min-width: 960px) {
-  .sm-hidden {
-    display: flex;
-  }
-}
-
-.title-xs {
-  font-size: 18px;
-  font-weight: 500;
-  margin: 20px 0;
-}
-
-.text-xs {
-  font-size: 18px;
-  font-weight: 300;
-}
