--- src/views/AboutUsPage/Support.jsx
@@ -1,155 +1,0 @@
-import PropTypes from "prop-types";
-
-const Support = ({ type }) => {
-  return (
-    <div>
-      {type === "support" && (
-        <svg
-          width="48"
-          height="51"
-          viewBox="0 0 48 51"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <g clipPath="url(#clip0_3766_1102)">
-            <path
-              d="M24.2784 29.75C12.5371 29.75 2.98438 38.4649 2.98438 49.1774C2.98438 49.6322 3.38837 50.0013 3.8875 50.0013H44.6692C45.1678 50.0013 45.5724 49.6322 45.5724 49.1774C45.5724 38.4649 36.0203 29.75 24.2784 29.75ZM4.81111 48.3535C5.28434 38.9307 13.8352 31.3977 24.2784 31.3977C34.7216 31.3977 43.273 38.9307 43.7456 48.3535H4.81111Z"
-              fill="#1D3967"
-              stroke="#1D3967"
-              strokeWidth="0.5"
-            />
-            <path
-              d="M23.9243 5.38281C17.1641 5.38281 11.6641 10.4002 11.6641 16.5671C11.6641 22.7341 17.1641 27.7514 23.9243 27.7514C30.6852 27.7514 36.1846 22.7341 36.1846 16.5671C36.1846 10.4002 30.6846 5.38281 23.9243 5.38281ZM23.9243 26.1037C18.16 26.1037 13.4703 21.8256 13.4703 16.5671C13.4703 11.3086 18.16 7.03055 23.9243 7.03055C29.6887 7.03055 34.3784 11.3086 34.3784 16.5671C34.3784 21.8256 29.6887 26.1037 23.9243 26.1037Z"
-              fill="#1D3967"
-              stroke="#1D3967"
-              strokeWidth="0.5"
-            />
-            <path
-              d="M9.03063 22.6925C9.52976 22.6925 9.93376 22.324 9.93376 21.8686V11.2067C9.93376 11.1539 9.92714 11.1028 9.9169 11.0529C12.3367 5.93444 17.7765 2.64719 23.9256 2.64719C30.1849 2.64719 35.7157 6.05362 38.0662 11.3341V21.8066C37.3088 23.5081 36.2088 25.0383 34.7849 26.3543C34.4345 26.6784 34.4381 27.1996 34.7933 27.5193C34.9691 27.678 35.1985 27.7571 35.4279 27.7571C35.6609 27.7571 35.8939 27.6753 36.0703 27.5122C37.5876 26.1099 38.7767 24.4891 39.6184 22.6931H40.253C43.9733 22.6931 47 19.932 47 16.5382C47 13.1444 43.9733 10.3833 40.253 10.3833H39.5799C36.8687 4.66952 30.7918 1 23.9256 1C17.0558 1 10.9784 4.66897 8.26839 10.3833H7.74698C4.02669 10.3833 1 13.1444 1 16.5382C1 19.932 4.02669 22.6931 7.74698 22.6931L9.03063 22.6925ZM45.1937 16.5376C45.1937 19.023 42.9769 21.0448 40.253 21.0448H39.8725V12.0305H40.253C42.9775 12.0305 45.1937 14.0523 45.1937 16.5376ZM2.80626 16.5376C2.80626 14.0523 5.02254 12.0305 7.74698 12.0305H8.1275V21.0448H7.74698C5.02254 21.0448 2.80626 19.023 2.80626 16.5376Z"
-              fill="#1D3967"
-              stroke="#1D3967"
-              strokeWidth="0.5"
-            />
-          </g>
-          <defs>
-            <clipPath id="clip0_3766_1102">
-              <rect width="48" height="51" fill="white" />
-            </clipPath>
-          </defs>
-        </svg>
-      )}
-
-      {type === "faq" && (
-        <svg
-          width="55"
-          height="61"
-          viewBox="0 0 55 61"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M25.2408 47.2803H2.34471C2.94284 37.4495 11.9957 29.6175 23.0386 29.6175C26.1788 29.6175 29.1949 30.2353 32.0043 31.453C32.5801 31.7018 33.27 31.4868 33.5482 30.9713C33.8257 30.4557 33.5852 29.8358 33.0102 29.5864C29.8847 28.2311 26.5295 27.5442 23.0379 27.5442C10.3354 27.5442 0 36.8621 0 48.3169C0 48.8891 0.517194 49.3535 1.15617 49.3535H25.24C25.879 49.3535 26.3962 48.8891 26.3962 48.3169C26.3962 47.7447 25.8797 47.2803 25.2408 47.2803Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M23.4972 25.1784C30.9406 25.1784 36.9959 19.7499 36.9959 13.0768C36.9959 6.40505 30.9406 0.976562 23.4972 0.976562C16.0568 0.976562 10.0039 6.40436 10.0039 13.0768C10.0039 19.7499 16.0568 25.1784 23.4972 25.1784ZM23.4972 3.05051C29.665 3.05051 34.6835 7.5488 34.6835 13.0775C34.6835 18.6069 29.665 23.1059 23.4972 23.1059C17.3317 23.1059 12.3162 18.6069 12.3162 13.0775C12.3162 7.54811 17.3317 3.05051 23.4972 3.05051Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M36.7749 41.7984C36.7749 37.1609 44.0624 37.1609 44.0624 41.7984C44.0624 45.1109 40.7499 44.4484 40.7499 48.4234M40.7499 53.7367L40.7632 53.7221"
-            stroke="#1D3967"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-          <path
-            d="M40.75 59.0234C48.068 59.0234 54 53.0914 54 45.7734C54 38.4555 48.068 32.5234 40.75 32.5234C33.432 32.5234 27.5 38.4555 27.5 45.7734C27.5 48.1863 28.1453 50.4507 29.2729 52.3984L28.1625 58.3609L34.125 57.2506C36.1384 58.4153 38.424 59.027 40.75 59.0234Z"
-            stroke="#1D3967"
-            strokeWidth="2"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </svg>
-      )}
-
-      {type === "forum" && (
-        <svg
-          width="66"
-          height="63"
-          viewBox="0 0 66 63"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M53.0653 30.5896C52.7625 30.4066 52.3831 30.3807 52.057 30.5219L39.0502 36.1412H30.7132C30.1307 36.1412 29.6592 36.5861 29.6592 37.1358V47.7445C29.6592 48.0323 29.7702 48.206 29.9438 48.448L38.9237 61.5623C39.1197 61.8362 39.4472 62.0006 39.7971 62.0006H44.9349C45.2891 62.0006 45.62 61.8329 45.8146 61.5531C46.01 61.2739 46.043 60.9205 45.9025 60.6135L40.8405 49.4996L52.0563 54.3319C52.1962 54.3922 52.3451 54.422 52.4941 54.422C52.6929 54.422 52.8911 54.369 53.0639 54.2642C53.3654 54.0812 53.5481 53.7663 53.5481 53.4275V31.427C53.5495 31.0882 53.3668 30.7726 53.0653 30.5896ZM31.7671 38.131H38.7937C38.7937 38.2404 38.7937 38.3459 38.7937 38.4626V46.7506H31.7671V38.131ZM40.3571 60.0115L32.2892 48.7397H38.2133C38.2239 48.7716 38.2309 48.8041 38.2456 48.8359L43.3357 60.0115H40.3571ZM51.4415 51.8799L40.9017 47.3387V37.5296L51.4415 32.9759V51.8799Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M58.5384 38.634C58.7499 38.634 58.9635 38.5743 59.149 38.4496L63.769 35.3459C64.2432 35.0277 64.3536 34.4064 64.0163 33.9588C63.6797 33.5119 63.0206 33.4065 62.5463 33.7254L57.9264 36.8292C57.4521 37.1474 57.3411 37.7687 57.679 38.2162C57.8849 38.4881 58.2088 38.634 58.5384 38.634Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M64.4477 40.7832H58.8264C58.2439 40.7832 57.7725 41.2281 57.7725 41.7778C57.7725 42.3274 58.2439 42.7723 58.8264 42.7723H64.4477C65.0295 42.7723 65.5017 42.3274 65.5017 41.7778C65.5017 41.2281 65.0295 40.7832 64.4477 40.7832Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M63.7686 48.2956L59.1487 45.1918C58.6758 44.8742 58.0167 44.9777 57.6787 45.4252C57.3415 45.8728 57.4518 46.4941 57.9261 46.8123L62.546 49.916C62.7315 50.0407 62.9444 50.1004 63.1566 50.1004C63.4862 50.1004 63.8108 49.9552 64.016 49.6826C64.3533 49.2358 64.2429 48.6145 63.7686 48.2956Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M21.4466 28.5301C24.3015 28.5301 27.0432 29.1209 29.5953 30.2859C30.1195 30.5259 30.7497 30.3184 31.0027 29.8237C31.2564 29.3291 31.0371 28.7344 30.513 28.4957C27.6721 27.1988 24.6219 26.541 21.4459 26.541C9.89564 26.541 0.499023 35.4583 0.499023 46.4184C0.499023 46.9681 0.970505 47.413 1.55301 47.413H23.6867C24.2692 47.413 24.7406 46.9681 24.7406 46.4184C24.7406 45.8687 24.2692 45.4238 23.6867 45.4238H2.6358C3.18106 36.0205 11.4099 28.5301 21.4466 28.5301Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M21.8625 24.1608C28.6312 24.1608 34.1379 18.9658 34.1379 12.5801C34.1379 6.19495 28.6312 1 21.8625 1C15.0974 1 9.59277 6.19428 9.59277 12.5801C9.59277 18.9658 15.0967 24.1608 21.8625 24.1608ZM21.8625 2.98913C27.469 2.98913 32.03 7.29162 32.03 12.5801C32.03 17.8692 27.469 22.1716 21.8625 22.1716C16.2596 22.1716 11.7007 17.8685 11.7007 12.5801C11.7007 7.29162 16.2596 2.98913 21.8625 2.98913Z"
-            fill="#1D3967"
-          />
-        </svg>
-      )}
-
-      {type === "contact" && (
-        <svg
-          width="40"
-          height="36"
-          viewBox="0 0 40 36"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M30.9433 22.2253C30.1761 21.4618 29.2455 21.0554 28.2548 21.0587C27.2721 21.062 26.3252 21.4672 25.5104 22.236L22.9612 24.6337C22.7545 24.5282 22.5478 24.4303 22.3491 24.3324C22.0629 24.1968 21.7925 24.0688 21.5622 23.933C19.2119 22.515 17.0818 20.664 15.0451 18.2665C14.0587 17.0715 13.398 16.0649 12.9211 15.0426C13.5821 14.4716 14.1954 13.8779 14.7924 13.3071C15.0183 13.0939 15.2443 12.8732 15.4702 12.6601C17.1643 11.0616 17.1855 8.99856 15.524 7.41139L13.3641 5.34807C13.1188 5.11377 12.8656 4.87191 12.6284 4.63001C12.1539 4.16136 11.6555 3.67762 11.1409 3.22428C10.3733 2.49874 9.45052 2.11503 8.47579 2.11831C7.50106 2.12159 6.55431 2.51159 5.74788 3.24242C5.73982 3.25003 5.73982 3.25003 5.73175 3.25764L2.98859 5.86837C1.95597 6.84267 1.36059 8.02787 1.21867 9.40115C1.0042 11.6166 1.67014 13.6773 2.18361 15.0257C3.44393 18.3359 5.3459 21.4012 8.19638 24.6833C11.6551 28.6384 15.8335 31.7568 20.6206 33.9478C22.4498 34.7684 24.8925 35.7386 27.6313 35.8962C27.799 35.9033 27.9747 35.9102 28.1344 35.9097C29.98 35.9035 31.5365 35.2687 32.7637 34.0132C32.7719 33.998 32.7879 33.9903 32.7961 33.9751C33.2165 33.4959 33.7003 33.0619 34.2084 32.5976C34.5551 32.2854 34.91 31.9581 35.257 31.6156C36.056 30.8317 36.4808 29.9202 36.4904 28.9872C36.5 28.0467 36.0859 27.1455 35.2867 26.3898L30.9433 22.2253ZM33.7216 30.2025C33.7136 30.2025 33.7135 30.2101 33.7216 30.2025C33.4067 30.5221 33.0842 30.8114 32.7374 31.1311C32.2132 31.6031 31.6809 32.0979 31.1798 32.6533C30.3645 33.4751 29.4097 33.8652 28.1634 33.8694C28.0435 33.8698 27.9157 33.8702 27.7959 33.863C25.4245 33.7269 23.2284 32.8545 21.5822 32.1091C17.0814 30.0461 13.1409 27.1089 9.87951 23.3807C7.18704 20.2725 5.39497 17.3964 4.22028 14.3059C3.49662 12.4197 3.24007 10.9492 3.37415 9.56076C3.46315 8.67307 3.83023 7.93613 4.49983 7.30436L7.25081 4.70877C7.64588 4.35855 8.06329 4.16753 8.47076 4.16616C8.97411 4.16447 9.37862 4.45132 9.6318 4.69318C9.63971 4.70073 9.64762 4.70829 9.65553 4.71585C10.1385 5.14655 10.5973 5.59249 11.0797 6.0687C11.3249 6.31058 11.5781 6.55243 11.8312 6.80187L13.9911 8.8652C14.8298 9.66634 14.8221 10.4097 13.967 11.2165C13.733 11.4372 13.5071 11.6579 13.2732 11.8711C12.5954 12.5257 11.95 13.1346 11.2488 13.7285C11.2327 13.7438 11.2166 13.7514 11.2085 13.7666C10.5147 14.4212 10.636 15.0579 10.7754 15.4898C10.7831 15.5125 10.7909 15.5352 10.7986 15.5579C11.3525 16.8606 12.1389 18.0867 13.3383 19.5465L13.3462 19.554C15.5249 22.1254 17.8292 24.1276 20.3781 25.6587C20.7036 25.8548 21.0376 26.013 21.3556 26.1636C21.6418 26.2991 21.9122 26.4272 22.1425 26.5629C22.1743 26.578 22.206 26.6006 22.2378 26.6157C22.5081 26.7437 22.7632 26.8035 23.0268 26.8027C23.6899 26.8004 24.1095 26.4046 24.2466 26.2752L27.0057 23.672C27.2799 23.4132 27.7146 23.1008 28.2179 23.0991C28.7133 23.0974 29.1177 23.3919 29.3627 23.6489C29.3707 23.6565 29.3707 23.6565 29.3786 23.664L33.738 27.8285C34.553 28.5994 34.5448 29.3958 33.7216 30.2025Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M21.6273 8.57993C23.7172 8.90664 25.6091 9.84082 27.1124 11.2769C28.6157 12.713 29.5799 14.515 29.919 16.5012C30.0018 17.0015 30.4536 17.3489 30.973 17.3471C31.0369 17.3469 31.0929 17.3392 31.1569 17.3314C31.7491 17.2383 32.146 16.7061 32.0559 16.1451C31.6491 13.742 30.4731 11.5539 28.6613 9.82297C26.8494 8.09209 24.5521 6.96206 22.0235 6.56097C21.4332 6.47194 20.8781 6.84547 20.7685 7.39955C20.659 7.95363 21.0369 8.49089 21.6273 8.57993Z"
-            fill="#1D3967"
-          />
-          <path
-            d="M38.9072 15.8523C38.2368 11.8951 36.3082 8.2987 33.3175 5.44161C30.3268 2.58452 26.5507 0.7313 22.387 0.0702356C21.8047 -0.0264116 21.2495 0.354708 21.14 0.908785C21.0384 1.47042 21.4245 1.99249 22.0147 2.08911C25.7318 2.67583 29.1103 4.34076 31.7846 6.88793C34.4589 9.44268 36.1836 12.6605 36.7785 16.1932C36.8613 16.6935 37.3131 17.0409 37.8325 17.0391C37.8964 17.0389 37.9524 17.0311 38.0164 17.0233C38.6005 16.9379 39.0054 16.4056 38.9072 15.8523Z"
-            fill="#1D3967"
-          />
-        </svg>
-      )}
-
-      {type === "whatsapp" && (
-        <svg
-          width="38"
-          height="38"
-          viewBox="0 0 38 38"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path
-            d="M18.9999 0C15.6735 0.000872649 12.4056 0.874726 9.52286 2.53422C6.64008 4.19372 4.24339 6.58073 2.57231 9.45667C0.901227 12.3326 0.0143004 15.5968 0.000171489 18.9229C-0.0139575 22.249 0.845207 25.5206 2.49179 28.4106L0.78085 34.3982C0.669188 34.7892 0.664098 35.2028 0.76611 35.5964C0.868121 35.9899 1.07352 36.349 1.36103 36.6365C1.64853 36.924 2.00769 37.1293 2.40127 37.2313C2.79485 37.3333 3.20855 37.3282 3.59949 37.2165L9.58769 35.5059C12.1158 36.9471 14.9412 37.7885 17.846 37.965C20.7508 38.1415 23.6574 37.6485 26.3414 36.524C29.0255 35.3995 31.4154 33.6736 33.3268 31.4793C35.2382 29.285 36.6201 26.681 37.3657 23.8682C38.1113 21.0554 38.2009 18.1089 37.6274 15.256C37.054 12.4031 35.8329 9.72002 34.0582 7.41372C32.2836 5.10743 30.0029 3.23954 27.3921 1.95414C24.7812 0.668738 21.91 0.000168607 18.9999 0ZM18.9999 36.4774C15.8605 36.4796 12.7787 35.6351 10.079 34.0329C9.99002 33.9799 9.8911 33.9457 9.78837 33.9323C9.68563 33.919 9.58126 33.9268 9.48167 33.9554L3.18187 35.7551C3.05156 35.7923 2.91368 35.794 2.78251 35.76C2.65133 35.7259 2.53164 35.6575 2.43583 35.5617C2.34002 35.4658 2.27157 35.3461 2.23758 35.215C2.20359 35.0838 2.2053 34.9459 2.24251 34.8156L4.04218 28.5164C4.07064 28.4168 4.07845 28.3125 4.06512 28.2098C4.05179 28.107 4.0176 28.0081 3.96467 27.9191C1.99586 24.6023 1.17937 20.7276 1.64239 16.8984C2.10541 13.0693 3.82194 9.50077 6.52458 6.74879C9.22723 3.9968 12.7642 2.21588 16.5846 1.68345C20.4049 1.15101 24.2941 1.89697 27.6463 3.8051C30.9985 5.71324 33.6255 8.67641 35.118 12.233C36.6105 15.7896 36.8847 19.74 35.8979 23.4687C34.9111 27.1973 32.7188 30.495 29.6624 32.848C26.606 35.2009 22.8571 36.477 18.9999 36.4774ZM28.5971 22.5216L24.7097 20.3005C24.358 20.0997 23.9593 19.9962 23.5543 20.0005C23.1494 20.0048 22.7529 20.1168 22.4056 20.325L19.7823 21.8988C18.1943 21.0943 16.9042 19.8043 16.0997 18.2163L17.6737 15.5932C17.8818 15.2458 17.9937 14.8494 17.998 14.4445C18.0023 14.0396 17.8987 13.6408 17.698 13.2892L15.4769 9.40206C15.2955 9.0856 15.0339 8.82258 14.7184 8.63952C14.4029 8.45646 14.0447 8.35984 13.6799 8.35941C12.0714 8.3598 10.5287 8.99733 9.38918 10.1325C8.24971 11.2676 7.60635 12.8079 7.59992 14.4162C7.59695 16.5138 8.00756 18.5915 8.80828 20.5303C9.60899 22.4691 10.7841 24.2311 12.2664 25.7154C13.7487 27.1998 15.5091 28.3773 17.4469 29.1808C19.3847 29.9842 21.4619 30.3978 23.5596 30.3979H23.5825C25.1909 30.3915 26.7313 29.7482 27.8666 28.6088C29.0018 27.4694 29.6394 25.9267 29.6398 24.3183C29.6394 23.9536 29.5428 23.5954 29.3597 23.2799C29.1766 22.9645 28.9136 22.7029 28.5971 22.5216ZM23.5804 28.878H23.5598C21.6618 28.878 19.7825 28.5038 18.0292 27.7769C16.276 27.05 14.6832 25.9846 13.3421 24.6417C12.0009 23.2988 10.9377 21.7046 10.2132 19.9505C9.48877 18.1963 9.11724 16.3165 9.11991 14.4187C9.12514 13.2128 9.60789 12.0582 10.4625 11.2074C11.3171 10.3567 12.474 9.87914 13.6799 9.87931C13.7768 9.87942 13.8719 9.90508 13.9557 9.95367C14.0395 10.0023 14.109 10.0721 14.1572 10.1561L16.3783 14.0434C16.4452 14.1606 16.4797 14.2935 16.4783 14.4285C16.4769 14.5634 16.4396 14.6956 16.3703 14.8114L14.5862 17.7843C14.5231 17.8894 14.4864 18.0082 14.4792 18.1306C14.4719 18.2529 14.4945 18.3752 14.5448 18.487C15.5367 20.6946 17.304 22.4617 19.5117 23.4535C19.6235 23.5038 19.7458 23.5263 19.8682 23.5191C19.9906 23.5119 20.1094 23.4752 20.2145 23.4121L23.1874 21.6281C23.3033 21.5587 23.4354 21.5214 23.5704 21.52C23.7054 21.5185 23.8383 21.553 23.9556 21.6199L27.843 23.8412C27.927 23.8894 27.9969 23.9588 28.0455 24.0426C28.0941 24.1264 28.1197 24.2215 28.1198 24.3183C28.12 25.5241 27.6424 26.6809 26.7917 27.5354C25.9409 28.39 24.7862 28.8727 23.5804 28.878Z"
-            fill="#1D3967"
-          />
-        </svg>
-      )}
-    </div>
-  );
-};
-
-export default Support;
-
-Support.propTypes = {
-  type: PropTypes.oneOf(["support", "faq", "forum", "whatsapp"]).isRequired,
-};
